Currency converter changes
