+++ .idea/misc.xml
@@ -3,10 +3,22 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_categories.xml" value="0.2269021739130435" />
         <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_convert_currency.xml" value="0.2269021739130435" />
-        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_home.xml" value="0.1" />
-        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_login.xml" value="0.1" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_home.xml" value="0.5136842105263157" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_login.xml" value="0.264" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_main.xml" value="0.2269021739130435" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_new_record.xml" value="0.13020833333333334" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/activity_sign_up.xml" value="0.13020833333333334" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/categories_item.xml" value="0.1" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/home_acc_adp.xml" value="0.536" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/home_expanses_adp.xml" value="0.1" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/input.xml" value="0.2269021739130435" />
         <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/loading.xml" value="0.2341485507246377" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/logo.xml" value="0.2269021739130435" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/nav_header.xml" value="0.2269021739130435" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/layout/user_category_input_layout.xml" value="0.2269021739130435" />
+        <entry key="..\:/Mokslai/Ketvirtas_kursas(2)/app_dev/App-dev-money-tracking/app/src/main/res/menu/drawer_menu.xml" value="0.4" />
         <entry key="..\:/Projects/App_dev_money_tracking/app/src/main/res/layout/activity_new_record.xml" value="0.22" />
         <entry key="..\:/Projects/App_dev_money_tracking/app/src/main/res/menu/drawer_menu.xml" value="0.24739583333333334" />
       </map>
