+++ app/src/main/res/layout/activity_home.xml
@@ -33,7 +33,7 @@
             <androidx.cardview.widget.CardView
                 android:id="@+id/cardView_home1"
                 android:layout_width="0dp"
-                android:layout_height="160dp"
+                android:layout_height="wrap_content"
                 android:layout_marginStart="5dp"
                 android:layout_marginTop="10dp"
                 android:layout_marginEnd="5dp"
@@ -47,7 +47,7 @@
 
                 <LinearLayout
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent"
+                    android:layout_height="wrap_content"
                     android:orientation="vertical">
 
                     <TextView
@@ -56,15 +56,50 @@
                         android:layout_height="wrap_content"
                         android:layout_marginLeft="10dp"
                         android:layout_marginTop="15dp"
-                        android:text="List of accounts"
+                        android:text="Balance"
                         android:textColor="#FFFFFF"
                         android:textSize="15sp" />
 
-                    <androidx.recyclerview.widget.RecyclerView
-                        android:id="@+id/Rec_view_list_of_acc"
+                    <LinearLayout
+                        android:id="@+id/lyt_lyn"
                         android:layout_width="match_parent"
-                        android:layout_height="match_parent"
-                        android:layout_margin="5dp" />
+                        android:layout_height="wrap_content"
+                        android:layout_marginBottom="5dp"
+                        android:orientation="horizontal">
+
+                        <EditText
+                            android:id="@+id/txt_e_balance"
+                            android:layout_width="191dp"
+                            android:layout_height="wrap_content"
+                            android:layout_marginStart="10dp"
+                            android:layout_marginEnd="5dp"
+                            android:alpha="0.8"
+                            android:background="@drawable/button_background"
+                            android:backgroundTint="#2BA5A1"
+                            android:ems="10"
+                            android:inputType="numberDecimal"
+                            android:maxLength="20"
+                            android:singleLine="false"
+                            android:text="10"
+                            android:textAlignment="center"
+                            android:textColor="#FFFFFF"
+                            app:layout_constraintBottom_toBottomOf="parent"
+                            app:layout_constraintEnd_toStartOf="@+id/guideline3"
+                            app:layout_constraintStart_toStartOf="parent"
+                            app:layout_constraintTop_toTopOf="parent" />
+
+                        <Button
+                            android:id="@+id/btn_adjust_b"
+                            android:layout_width="179dp"
+                            android:layout_height="36dp"
+                            android:layout_margin="5dp"
+                            android:layout_weight="1"
+                            android:background="@drawable/button_background"
+                            android:text="Button"
+                            android:textColor="#FFFFFF" />
+
+                    </LinearLayout>
+
                 </LinearLayout>
 
             </androidx.cardview.widget.CardView>
@@ -131,7 +166,6 @@
                             android:layout_marginBottom="5dp"
                             android:background="@drawable/button_background"
                             android:backgroundTint="#512DA8"
-                            android:onClick="On_show_more_click"
                             android:text="Show more"
                             android:textColor="#FFFFFF"
                             android:textSize="12sp" />
@@ -148,7 +182,6 @@
                             android:layout_weight="1"
                             android:background="@drawable/button_background"
                             android:backgroundTint="@color/primary"
-                            android:onClick="On_add_record_click"
                             android:text="Add record"
                             android:textColor="#FFFFFF"
                             android:textSize="12sp" />
