+++ app/src/main/res/menu/drawer_menu.xml
@@ -15,5 +15,9 @@
             android:id="@+id/nav_categories"
             android:icon="@drawable/ic_category"
             android:title="Categories" />
+        <item
+            android:id="@+id/nav_converter"
+            android:icon="@drawable/investment"
+            android:title="Currency converter" />
     </group>
 </menu>
