+++ app/src/main/res/values/strings.xml
@@ -14,6 +14,18 @@
         <item>HKD</item>
         <item>NZD</item>
     </string-array>
+    <string-array name="currency_names">
+        <item>EUR (Euro)</item>
+        <item>USD (United States Dollar)</item>
+        <item>JPY (Japanese Yen)</item>
+        <item>GBP (British Pound Sterling)</item>
+        <item>AUD (Australian Dollar)</item>
+        <item>CAD (Canadian Dollar)</item>
+        <item>CHF (Swiss Franc)</item>
+        <item>CNH (Chinese Yuan)</item>
+        <item>HKD (Hong Kong Dollar)</item>
+        <item>NZD (New Zealand Dollar)</item>
+    </string-array>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
